+++ solvers/CardiacEPSolver/CellModels/CourtemancheRamirezNattel98.cpp
@@ -187,8 +187,8 @@ void CourtemancheRamirezNattel98::v_Update(
     //  19  K_i  Potassium
     //  20  Ca_rel Calcium Rel
     //  21  Ca_up  Calcium up
-    int n = m_nq;
-    int i = 0;
+    size_t n = m_nq;
+    size_t i = 0;
     NekDouble alpha, beta;
     Vmath::Zero(n, outarray[0], 1);
 
@@ -627,7 +627,7 @@ void CourtemancheRamirezNattel98::v_SetInitialConditions()
     Vmath::Fill(m_nq, 1.488, m_cellSol[20], 1);
 }
 
-std::string CourtemancheRamirezNattel98::v_GetCellVarName(unsigned int idx)
+std::string CourtemancheRamirezNattel98::v_GetCellVarName(size_t idx)
 {
     switch (idx)
     {
